+++ src/js/NativeRNSentry.ts
@@ -44,6 +44,8 @@ export interface Spec extends TurboModule {
   fetchNativePackageName(): string | undefined | null;
   fetchNativeStackFramesBy(instructionsAddr: number[]): NativeStackFrames | undefined | null;
   initNativeReactNavigationNewFrameTracking(): Promise<void>;
+  captureReplay(isHardCrash: boolean): Promise<string | undefined | null>;
+  getCurrentReplayId(): string | undefined | null;
 }
 
 export type NativeStackFrame = {